pyproject.toml
[tool.poetry]
name = "discord-reacji-channeler"
version = "0.1.0"
description = "ABC"
authors = ["repl.it user <replituser@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "3.6.2"
astroid = "2.4.2"
async-timeout = "3.0.1"
attrs = "20.2.0"
chardet = "3.0.4"
colorama = "0.4.3"
DateTime = "4.3"
"discord.py" = "^1.6.0"
discord-py-slash-command = "^1.0.9"
idna = "2.10"
isort = "5.5.1"
lazy-object-proxy = "1.4.3"
mccabe = "0.6.1"
multidict = "4.7.6"
pylint = "2.6.0"
python-dateutil = "2.8.1"
python-dotenv = "0.14.0"
pytz = "2020.1"
six = "1.15.0"
toml = "0.10.1"
typed-ast = "1.4.1"
typing-extensions = "3.7.4.3"
websockets = "6.0"
wrapt = "1.12.1"
yarl = "1.5.1"
"zope.interface" = "^5.1.0"
mojimoji = "0.0.11"
flask = "^1.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
